// ui-frontend-portal/src/app/sensors/model/WeatherStation.ts
/**
 * Domain Persistence API
 * This API is the common interface to access the persistence domain for all the platform components
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface WeatherStation {
  id?: number;

  cityId?: number;

  name?: string;

  description?: number;

  lat?: number;

  lng?: number;

}