// ui-frontend-portal/src/app/sensors/model/WeatherReading.ts
/**
 * Domain Persistence API
 * This API is the common interface to access the persistence domain for all the platform components
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface WeatherReading {
  weatherStationId?: number;

  timestamp?: Date;

  amountPrecipitation?: number;

  dewpointTemperature?: number;

  temperature?: number;

  reducedAirPressure?: number;

  windForce?: number;

}